new commit

+++ src/Libraries/Nop.Services/Catalog/ProductTagService.cs
@@ -2,7 +2,6 @@
 using System.Collections.Generic;
 using System.Linq;
 using System.Threading.Tasks;
-using DocumentFormat.OpenXml.Office2010.Excel;
 using Nop.Core;
 using Nop.Core.Caching;
 using Nop.Core.Domain.Catalog;
@@ -72,7 +71,8 @@ namespace Nop.Services.Catalog
             var mappingRecord = await _productProductTagMappingRepository.Table
                 .FirstOrDefaultAsync(pptm => pptm.ProductId == productId && pptm.ProductTagId == productTagId);
 
-            ArgumentNullException.ThrowIfNull(mappingRecord); //izmenjeno remove if
+            if (mappingRecord is null)
+                throw new Exception("Mapping record not found");
 
             await _productProductTagMappingRepository.DeleteAsync(mappingRecord);
         }
